chore: Update npm dependencies
This commit updates the npm dependencies in the `package.json` file. It adds new dependencies such as `@chakra-ui/icons`, `@chakra-ui/react`, `@emotion/react`, `@emotion/styled`, `framer-motion`, and `react-router-dom`. These updates are necessary to ensure compatibility with the latest versions of the libraries and to take advantage of any bug fixes or new features they provide.